Implement transformation shaders and component management system
- Added transformation functions in GLSL for translation, scaling, and rotation.
- Introduced a new `ComponentStorage` structure for managing components with thread-safe access.
- Implemented a `ComponentRegistry` to register and retrieve component storages by type.
- Enhanced the `Scene` struct to manage entities and their components effectively.
- Integrated GLTF loading functionality to parse and instantiate 3D models into the scene.
- Added support for applying transformations recursively to child entities in the scene.
- Fix: get_dirty did not retrieve odd bits
- Fix: not all transform data was iterated over to send to gpu

+++ src/gpu_manager/gpu_vec.rs
@@ -63,7 +63,7 @@ where
         }
         let buf = gpu
             .sub_alloc(self.usage)
-            .allocate_slice(self.data.len() as u64 - self.size as u64)
+            .allocate_slice((self.data.len() - self.size) as u64)
             .unwrap();
         {
             let mut write = buf.write().unwrap();
